Add files via upload
A telecommunications provider has segmented its customer base by service usage patterns, categorizing the customers into four groups. If demographic data can be used to predict group membership, the company can customize offers for individual prospective customers. It is a classification problem. That is, given the dataset, with predefined labels, we need to build a model to be used to predict class of a new or unknown case.

It focuses on using demographic data, such as region, age, and marital, to predict usage patterns.

The target field, called custcat, has four possible service categories that correspond to the four customer groups, as follows:

    Basic Service
    E-Service
    Plus Service
    Total Service

The objective is to build a classifier to predict the service category for unknown cases.